Add lookup-bench CLI and adapt to sshash-rs i64 orientation
- Add lookup-bench subcommand for isolated sshash streaming/point benchmarks
  with --no-locate flag to measure pure dictionary lookup time
- Update unitig_end_cache test and tiny_sshash_parity test for i8→i64
  kmer_orientation change in sshash-rs

+++ src/mapping/unitig_end_cache.rs
@@ -100,7 +100,7 @@ impl UnitigEndCache {
 mod tests {
     use super::*;
 
-    fn make_result(string_id: u64, kmer_id: u64, orientation: i8) -> LookupResult {
+    fn make_result(string_id: u64, kmer_id: u64, orientation: i64) -> LookupResult {
         LookupResult {
             kmer_id,
             kmer_id_in_string: 5,
